+++ actiontext/app/javascript/actiontext/attachment_upload.js
@@ -6,6 +6,7 @@ export class AttachmentUpload {
     this.element = element
     this.directUpload = new DirectUpload(file, this.directUploadUrl, this)
     this.file = file
+    this.dispatch("initialize")
   }
 
   start() {
@@ -15,7 +16,12 @@ export class AttachmentUpload {
     })
   }
 
+  directUploadWillCreateBlobWithXHR(xhr) {
+    this.dispatch("before-blob-request", { xhr })
+  }
+
   directUploadWillStoreFileWithXHR(xhr) {
+    this.dispatch("before-storage-request", { xhr })
     xhr.upload.addEventListener("progress", event => {
       // Scale upload progress to 0-90% range
       const progress = (event.loaded / event.total) * 90
@@ -91,6 +97,8 @@ export class AttachmentUpload {
 
   dispatch(name, detail = {}) {
     detail.attachment = this.attachment
+    detail.file = this.directUpload.file
+    detail.id = this.directUpload.id
     return dispatchEvent(this.element, `direct-upload:${name}`, { detail })
   }